Revert "Preserve L2 ROP time across cycle wrap"
This reverts commit 769f7c88e3b75c3ff253bd40c659b3781a3522ee.

+++ src/gpgpu-sim/l2cache.h
@@ -79,7 +79,7 @@ class memory_partition_unit {
 
   bool busy() const;
 
-  void cache_cycle(unsigned long long cycle);
+  void cache_cycle(unsigned cycle);
   void dram_cycle();
   void simple_dram_model_cycle();
 
@@ -169,7 +169,7 @@ class memory_sub_partition {
 
   bool busy() const;
 
-  void cache_cycle(unsigned long long cycle);
+  void cache_cycle(unsigned cycle);
 
   bool full() const;
   bool full(unsigned size) const;
